+++ src/exit-with-error-internal.ts
@@ -1,11 +1,9 @@
 import exitWithError from "./exit-with-error.js";
 import { getHelpCommand } from "./help/help.js";
-import type { ParsedCommand } from "./types/parsed";
 
-
-export function exitWithErrorInternal(message: string, parsed: ParsedCommand): never {
+export function exitWithErrorInternal(message: string, treePath: string[]): never {
     exitWithError(
         message
-        + `\nUse "${getHelpCommand(parsed)}" to get help on this command`
+        + `\nUse "${getHelpCommand(treePath)}" to get help on this command`
     )
 }